requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
google-auth==2.28.0
google-auth-oauthlib==1.2.0
requests==2.31.0
beautifulsoup4==4.12.2
html5lib==1.1
lxml==5.1.0
PyJWT==2.8.0 